+++ bingads/manifest.py
@@ -1,6 +1,5 @@
 import sys
-VERSION = '13.0.2'
-BULK_FORMAT_VERSION_5 = '5.0'
+VERSION = '13.0.3'
 BULK_FORMAT_VERSION_6 = '6.0'
 WORKING_NAME = 'BingAdsSDKPython'
 USER_AGENT = '{0} {1} {2}'.format(WORKING_NAME, VERSION, sys.version_info[0:3])